finished square/bit conversion templates fixed unsigned/signed error in rotated coordinates

+++ Board/GhostPredicates.h
@@ -7,7 +7,7 @@ class BIT_IS_GHOST;
 template<typename, size_t>
 class BIT2COORD;
 
-template<typename, size_t, size_t>
+template<typename, int, int>
 class COORD2BIT;
 
 // include template definitions inside header since "export" keyword is not supported by Visual C++
